version creation moved to a helper
